SmallTree(CAIC)	R Documentation

Example dataset for CAIC analyses

Description

SmallTree is an artificially created dataset used to illustrate methods in the CAIC package.

Usage


data(SmallTree)

Format

SmallTree contains two phylogenies stored in the ape package 'phylo' format. The first (diTree) is a 15 tip, fully bifurcating phylogeny. The second (polyTree) is based on diTree but has had three nodes collapsed to give three way polytomies. The are also two accompanying datasets (SmallTreeDat and SmallTreeDatNA), which contain data for each of the tips in the tree. They contain a column of tip names (tip), two continuous variables (weight and othervar), a column of the number of species in the tip group (nSpp) and two categorical variables (catX2 and catX3). SmallTreeDatNA differs only in having two lines for which no data is available.

One reason these comparatively trivial toy examples have been preserved as a dataset is that they are also used to provide a reference dataset back to analyses performed in the original CAIC and MacroCAIC programs for Mac Classic.

See Also

caic, macrocaic

[Package CAIC version 1.0 Index]