caic(CAIC)	R Documentation

Comparative analysis by independent contrasts.

Description

Calculate a linear model using phylogenetic independent contrasts using a formula interface.

Usage


caic(formula, data, phy, names.col, contr.type = "Crunch", stand.contr = TRUE, ref.var = NULL, node.depth = NULL, crunch.brlen = NULL, macroMinSize = 3, equal.branch.length = FALSE)

Arguments

formula 	 A model formula.

data 	 A data frame containing the variables specified in the model formula. The data frame can contain missing data for explanatory variables.

phy 	 An object of class 'phylo' giving the phylogenetic structure in the data.

names.col 	 A name specifying the column in 'data' that matches rows to tips in 'phy'.

contr.type 	 Specifies whether to calculate contrasts using the 'Crunch' or 'Brunch' algorithm. See details.

stand.contr 	 A logical flag indicating whether or not to standardize contrasts

ref.var 	 A reference variable present in the model that is used to specify the direction of calculation for polytomies. If null, this is assumed to be the first explanatory variable.

node.depth 	 A positive integer greater than 1 used to restrict the model to contrasts with a node depth less than or equal to the specified depth. Tips have a depth of 1.

crunch.brlen 	 The internal branch length used for calculating contrasts at a polytomy.

equal.branch.length 	 If set to 'TRUE' then all branch lengths are set to 2.

Details

These functions implement methods for modelling the relationship between variables that are phylogenetically non-independent. The method was first described by Felsenstein (1985) and subsequently extended to permit the use of phylogenies with polytomies by Pagel (1992). These methods were previously implemented in the Mac Classic computer programs CAIC, written by Andy Purvis, Andy Rambaut (Purvis and Rambaut, 1995) and updated by Nick Isaac and Paul-Michael Agapow.

The user provides a data frame and a phylogeny. These are matched using names.col to identify a column in the data frame that contains tip labels on the phylogeny. The formula specifies the model to be fitted and contrasts are calculated in those variables. The specified reference variable is used to ensure that contrasts for multivariate models are calculated in a consistent direction at each node. The function caic acts as a data preparation wrapper, finding the union between the phylogeny and data and reducing the dataset to the correct model and response matrices. The prepared data is then sent to the contrCalc engine, which is not intended to be called by users.

The resulting table of contrasts is then used to fit the specified model - note that the intercept is automatically dropped from the model if present, following REFHERE.

Missing data notes here...

The functions can be used to calculate either 'Crunch' or 'Brunch' style contrasts. More info here... Polytomies - more detail here The Mac Classic program CAIC used 1 for both 'Brunch' and 'Crunch' analyses and this the default.

Value

A list containing the following:

data 	 A list containing the internal node numbers, the contrasts, nodal values from the phylogeny ...EXTEND...

phy 	 A 'phylo' object describing the phylogeny used to calculate contrasts. This is the phylogeny in the original call reduced to those tips that are also named in the provided data but possibly including tips with missing data.

lm 	 An 'lm' object containing a regression model through the origin for the calculated contrasts

Author(s)

David Orme

References

Felsenstein, J. (1985). Phylogenies and the comparative method. Am. Nat. 125, 1–15 Pagel, M. D. (1992). A method for the analysis of comparative data. J. theor. Biol. 156, 431-442. Purvis, A. and Rambaut, A. (1995) Comparative analysis by independent contrasts (CAIC): an Apple Macintosh application for analysing comparative data. Computer Appl. Biosciences 11, 247-251.

See Also

caic.table for extracting a data table, plot.caic for contrast plots, summary.caic for a data and model summary.

Examples



data(SmallTree)
caicMod <- caic(othervar ~ weight, phy=diTree, data=SmallTreeDat, names.col=tip)

summary(caicMod)

# plot the contrasts
caicTab <- caic.table(caicMod)
plot(othervar ~ weight, data=caicTab)

# for the actual model diagnostics
plot(caicMod)


[Package CAIC version 1.0 Index]