caic.table(CAIC)	R Documentation

Extract a data frame from a CAIC object.

Description

A CAIC object contains a relatively complex list (obj$contrast.data) containing contrasts and other values. This function extracts this data to a data frame, optionally including information on all nodes in the phylogeny, not just those for which valid contrasts can be calculated. There is also an option to include nodal values for the internal nodes in the data frame.

Usage


caic.table(caicObj, validNodes = TRUE, nodalValues=FALSE)

Arguments

caicObj 	 A CAIC object produced by either caic() or macrocaic().

validNodes

nodalValues 	 ~~Describe validNodes here~~

Details

The function is also called internally by caic and macrocaic to provide a model frame for the 'lm' object.

Value

A data frame containing

Author(s)

David Orme

See Also

caic,macrocaic

Examples


data(SmallTree)
caicMod <- caic(othervar ~ weight, phy=diTree, data=SmallTreeDat, names.col=tip)
caic.table(caicMod)


[Package CAIC version 1.0 Index]