clade.members(CAIC)	R Documentation

Identify tips descended from a node

Description

Obtains a vector of the tips subtending from either one node or all nodes in a phylogeny.

Usage


clade.members(x, phyl, tip.labels = FALSE)

Arguments

x 	 A integer in the range minus number of internal of nodes to positive number of tips.

phyl 	 An object of class 'phylo'.

tips 	 A logical indicating whether to include external node membership in the list.

tip.labels 	 A logical flag indicating whether to return the node numbers of the tips or their tip labels.

Details

The function all.clades runs clade.members over each node in the phylogeny, possibly including the external nodes as indicated by the tips argument, and returns a list of vectors showing the members of the clade defined by each node.

Value

A numeric vector of external node (i.e. tip) numbers or a character vector of tip labels for a single internal node or, for all.clades, a list of such vector for all nodes in the phylogeny.

Author(s)

David Orme

See Also

clade.stats

Examples


data(SmallTree)
clade.members(22, diTree, tip.labels=TRUE)
all.clades(diTree, tip.labels=FALSE)

[Package CAIC version 1.0 Index]