node2tip(CAIC)	R Documentation

Calculate node depth in a phylogeny

Description

This function calculates node depth as the maximum number of nodes between each node in a phylogeny and the tips.

Usage


node2tip(phy)

Arguments

phy 	 A 'phylo' object.

Details

The tips are counted and so all tips have a node to tip distance of 1.

Value

A vector of the node to tip distances, named using the node numbers from the phylogeny edge matrix.

Author(s)

David Orme

Examples


data(SmallTree)
node2tip(diTree)

[Package CAIC version 1.0 Index]