summary.caic(CAIC)	R Documentation

Summarize a caic or macrocaic analysis

Description

Currently both these functions print a brief description of the datasets and then a summary of the resulting linear model from a caic or macrocaic analysis.

Usage


summary.caic(obj)

Arguments

obj 	 ~~Describe obj here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples




[Package CAIC version 1.0 Index]