write.caic(CAIC)	R Documentation

Writes a 'phylo' object to a CAIC phylogeny file

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


write.caic(phy, filebase, equal.brlen, charset)

Arguments

phy 	 A phylo object

filebase 	 A text string used as the file base name for the .phyl and .blen files

equal.brlen 	 If provided, all branches in the branch length file are set to this value. By default, the branch lengths in the phylo object are used or, if the phylogeny has no branch lengths, equal branch lengths of 2 are used.

charset 	 Can be used to extend the number of coding characters available for extremely polytomous trees. Defaults to LETTERS.

Value

Returns NULL.

Author(s)

David Orme

References

~put references to the literature/web site here ~

See Also

caic.label

Examples


data(bird.families)
write.caic(bird.families, filebase="bird.families")

[Package CAIC version 1.0 Index]